+++ src/data/systems/broken-compass/components/DiamondSkill.svelte
@@ -56,6 +56,7 @@
 
 </script>
 
+{#if skill !== undefined}
 <div class="flex items-center">
   {#if $editMode}
     <button type="button" on:click={skillDecrement} class="btn-icon btn-icon-sm variant-filled mr-1">-</button>
@@ -74,3 +75,4 @@
     <button type="button" on:click={skillIncrement} class="btn-icon btn-icon-sm variant-filled ml-1">+</button>
   {/if}
 </div>
+{/if}