+++ src/routes/login/+page.svelte
@@ -57,53 +57,17 @@
     <meta name="description" content="Log In or Register">
 </svelte:head>
 
-<div class="flex flex-col justify-center items-center">
+<div class="flex flex-col items-center my-3">
 
     {#if $currentUser}
-        <p>Signed in as {$currentUser.username}</p>
+        <h1 class="h1 m-auto my-3 flex">Signed in as {$currentUser.email}</h1>
     {:else}
-        <div class="tabs">
-            <a on:click|preventDefault={() => isLogin = true} href="/" class="tab tab-lg {isLogin  === true ? 'tab-active' : ''}">Login</a> 
-            <a on:click|preventDefault={() => isLogin = false} href="/" class="tab tab-lg {isLogin  !== true ? 'tab-active' : ''}">Register</a>
-        </div>
-        
-        <div class="m-auto mt-3">
+    
+    
+    <h1 class="h1 m-auto my-3 flex">Login</h1>
 
-            <form on:submit|preventDefault class="flex-col">
+    <hr class="w-full" />
 
-                <div>
-                    <input
-                        class="input input-lg {isError ? " input-bordered input-error " : ""}"
-                        type="text"
-                        name="username"
-                        placeholder="username"
-                        bind:value={username} />
-                    <input
-                        class="input input-lg {isError ? " input-bordered input-error " : ""}"
-                        type="password"
-                        name="password"
-                        placeholder="password"
-                        bind:value={password} />
-                </div>
-
-        
-                {#if isError}
-                <div class="alert alert-error mt-3">
-                    <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current shrink-0 h-6 w-6" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
-                    <span>{errorText}</span>
-                </div>
-                {/if}
-
-                <div class="mt-3">
-                    {#if isLogin === true}
-                        <button class="btn" on:click={login}>Login</button>
-                    {:else}
-                        <button class="btn" on:click={signUp}>Sign Up</button>
-                    {/if}
-                </div>
-            </form>
-        </div>
-
-        <button class="btn variant-filled" on:click={googleLogin}><Icon src={BsGoogle} /></button>
+    <button class="btn variant-filled mt-3" on:click={googleLogin}><Icon src={BsGoogle} />oogle</button>
     {/if}
 </div>