+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - Feature: Adds primitive DnD support
 - Feature: Adds datatable field type
 - UI/UX: Design fixes
+- Feature: Adds Meta
+- Feature: Login simplified to Google-only for now
+- Feature: big character sheet refactoring
 
 ### 0.1.7
 - Feature: Adds logo